requirements.txt
Django==1.11.4
djangorestframework==3.6.4
django-rest-framework==0.1.0
django-filter==1.0.4
